base/encrypt and decrypt/aes.go

package algorithm

import (
	"crypto/aes"
	"crypto/cipher"
)

const (
	TopSecretKey = "#@sand@shanghai!"
)

type sandAES struct{}

func NewAESEnDecrypter() EnDecrypter {
	return &sandAES{}
}

func (this *sandAES) Encrypt(key, data []byte) ([]byte, error) {
	block, err := aes.NewCipher(key)
	if err != nil {
		return nil, err
	}
	blockSize := block.BlockSize()
	data = PKCS5Padding(data, blockSize)
	blockMode := cipher.NewCBCEncrypter(block, key[:blockSize])
	crypted := make([]byte, len(data))
	blockMode.CryptBlocks(crypted, data)
	return crypted, nil
}

func (this *sandAES) Decrypt(key, data []byte) ([]byte, error) {
	block, err := aes.NewCipher(key)
	if err != nil {
		return nil, err
	}
	blockSize := block.BlockSize()
	blockMode := cipher.NewCBCDecrypter(block, key[:blockSize])
	origData := make([]byte, len(data))
	blockMode.CryptBlocks(origData, data)
	origData = PKCS5UnPadding(origData)
	return origData, nil
}
